adds hello world app with common.js config for custom app
