+++ server.js
@@ -1,0 +1,45 @@
+require('isomorphic-fetch');
+import Koa from 'koa';
+import next from 'next';
+import createShopifyAuth, { verifyRequest } from '@shopify/koa-shopify-auth'
+import dotenv from 'dotenv';
+import session from 'koa-session';
+dotenv.config();
+
+const port = parseInt(process.env.PORT, 10) || 3000;
+const dev = process.env.NODE_ENV !== 'production';
+const app = next({ dev });
+const handle = app.getRequestHandler();
+
+const { SHOPIFY_API_SECRET_KEY, SHOPIFY_API_KEY } = process.env;
+
+app.prepare().then(() => {
+  const server = new Koa();
+  server.use(session(server));
+  server.keys = [SHOPIFY_API_SECRET_KEY];
+
+  server.use(
+    createShopifyAuth({
+      apiKey: SHOPIFY_API_KEY,
+      secret: SHOPIFY_API_SECRET_KEY,
+      scopes: ['read_products'],
+      afterAuth(ctx) {
+        const { shop, accessToken } = ctx.session;
+
+        ctx.redirect('/');
+      },
+    }),
+  );
+
+  server.use(verifyRequest());
+  server.use(async (ctx) => {
+    await handle(ctx.req, ctx.res);
+    ctx.respond = false;
+    ctx.res.statusCode = 200;
+    return
+  });
+
+  server.listen(port, () => {
+    console.log(`> Ready on http://localhost:${port}`);
+  });
+});
